Security group changes.
This is the security group changes based on the input from Matt.

Allows LDAP to OID from the other weblogic domains.

Allows the admin/managed ports down from the ELB into the domain servers.

Allows DB connections from the servers that need to get to the DB with split for Delius DB and OID DB.

SSH is allowed from bastion to almost everywhere.

Did not include JMX port as that can be delivered via SSH tunneling. To be discussed with security.