Feature/shared vpc security groups (#11)
* Weblogic module, more weblogic.

This adds a module for weblogic setup that gives a single instance of an admin node and a managed node to allow us to get the Ansible working for both.

Currently commented out two of the tiers as we don't have the subnet space, need to re-allocate the subnets later.

* Security group changes.

This is the security group changes based on the input from Matt.

Allows LDAP to OID from the other weblogic domains.

Allows the admin/managed ports down from the ELB into the domain servers.

Allows DB connections from the servers that need to get to the DB with split for Delius DB and OID DB.

SSH is allowed from bastion to almost everywhere.

Did not include JMX port as that can be delivered via SSH tunneling. To be discussed with security.

* Changed the role from 'admin' to 'terraform' enable Jenkins slave to run code.

* removed 'apply' to eanble usage of the plan file in terragrunt apply

* namespaced rules with identical names
used the first letter of each part of group

* Commented out rule as it's a WIP that requires either
security_group_id or CIDR for completion.

* added id to the data lookup

* fixing refactor oversights

* renamed file to match resource name

* Put all security group rule attributes into a consistent order.

* Added missing attribute

* Duplicate attribute removed

* Put project remote states into name spaced directory in S3 bucket.

* removed bastion security group as it's now in the shared vpc terraform project.

* renamed properties files to drop suffix and include 'delius-core-' so that they tie in with same env names in shared VPC terraform project

* Updated properties and vars files to remove duplicated values and align with VPC values.

* Security groups migrated over to use shared VPC remote state etc

* Added security group id outputs

* Updated security group rules to use bastion cidrs output from shared vpc remote state
re-enabled the remaining rules

* Refactored the application terraform to use the shared vpc and shared security groups remote state.
Updated the delius db to use these - tested access from bastion WIP

* Made changes to enable both db servers to deploy in shared VPC

* Tidying up unused variables as those values are now supplied by the shared vpc remote state.
Renamed the tfvars files in line and updated where they are referenced.

* removed deprecated variable

* formatted

* Module for weblogic servers refactored for use with shared vpc
Outputs added
errors fixed

* All weblogic tiers migrated to shared vpc compatibility.
split into individual files as now quite long.

* Removed keys directory as deployer key managed in shared VPC.
KMS key managed in application dir until added to shared vpc code.