Changed source of "high_availability_count" from key in db_size map to stand alone variable.
During development of the variable HA quantity solution the location of the variable was moved from TF VARs to Ansible Vars to ensure one single source.

This was over looked and currently the value is using the default of 2 when it doesn't find the value in the db_size tf_var map.

Solution:
Make in to a standalone variable, and pass in to the env as TF_VAR.

Use the value from the Ansible vars in env configs which is already used to determine how many HAs to build.